@@ -13,16 +13,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.1.0",
+    "@angular/animations": "^18.2.14",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.1.0",
-    "@angular/compiler": "^17.1.0",
-    "@angular/core": "^17.1.0",
-    "@angular/forms": "^17.1.0",
+    "@angular/common": "^18.2.14",
+    "@angular/compiler": "^18.2.14",
+    "@angular/core": "^18.2.14",
+    "@angular/forms": "^18.2.14",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.1.0",
-    "@angular/platform-browser-dynamic": "^17.1.0",
-    "@angular/router": "^17.1.0",
+    "@angular/platform-browser": "^18.2.14",
+    "@angular/platform-browser-dynamic": "^18.2.14",
+    "@angular/router": "^18.2.14",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.3"
@@ -30,7 +30,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.1.1",
     "@angular/cli": "^17.3.16",
-    "@angular/compiler-cli": "^17.1.0",
+    "@angular/compiler-cli": "^18.2.14",
     "@types/chrome": "^0.0.268",
     "@types/jasmine": "~5.1.0",
     "jasmine-core": "~5.1.0",
