@@ -14,7 +14,7 @@ import {provideAnimationsAsync} from '@angular/platform-browser/animations/async
 import {MatIconModule, MatIconRegistry} from "@angular/material/icon";
 import {MatButtonModule} from "@angular/material/button";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
-import {HttpClientModule} from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import {XIconComponent} from "../ui/marker/x-marker/x-icon/x-icon.component";
 import {XMarkerComponent} from "../ui/marker/x-marker/x-marker.component";
 import {MarkerTooltipComponent} from "../ui/marker/marker-tooltip/marker-tooltip.component";
@@ -25,35 +25,28 @@ import {GroupingMarkerComponent} from "../ui/marker/grouping-marker/grouping-mar
 import {MarkerComponent} from "../ui/marker/marker-component";
 
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    TimelineComponent,
-    StreamTimelineComponent,
-    PanelComponent,
-    NewPanelComponent,
-    XIconComponent,
-    PanelEmissionsTimelineComponent,
-    XMarkerComponent,
-    MarkerTooltipComponent,
-    MarkerTooltipDirective,
-    VerticalSplitPaneComponent,
-    TagComponent,
-    GroupingMarkerComponent,
-    MarkerComponent
-  ],
-  imports: [
-    CommonModule,
-    BrowserModule,
-    RouterModule.forRoot(routes),
-    BrowserAnimationsModule,
-    MatButtonModule,
-    MatIconModule,
-    HttpClientModule
-  ],
-  bootstrap: [AppComponent],
-  providers: [ColorModifier, provideAnimationsAsync()]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        TimelineComponent,
+        StreamTimelineComponent,
+        PanelComponent,
+        NewPanelComponent,
+        XIconComponent,
+        PanelEmissionsTimelineComponent,
+        XMarkerComponent,
+        MarkerTooltipComponent,
+        MarkerTooltipDirective,
+        VerticalSplitPaneComponent,
+        TagComponent,
+        GroupingMarkerComponent,
+        MarkerComponent
+    ],
+    bootstrap: [AppComponent], imports: [CommonModule,
+        BrowserModule,
+        RouterModule.forRoot(routes),
+        BrowserAnimationsModule,
+        MatButtonModule,
+        MatIconModule], providers: [ColorModifier, provideAnimationsAsync(), provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {
   private domSanitizer: DomSanitizer;
 
